+++ src/help/help_features.rs
@@ -8,12 +8,6 @@ pub fn list() -> Vec<(&'static str, &'static str)> {
     #[cfg(not(any(target_family = "windows", target_os = "android")))]
     features.push(("permissions", "allow showing file mode, owner and group"));
 
-    #[cfg(feature = "client-server")]
-    features.push((
-        "client-server",
-        "see https://github.com/Canop/broot/blob/master/client-server.md",
-    ));
-
     #[cfg(feature = "clipboard")]
     features.push((
         "clipboard",